test: add controller test for create missions
